+++ src/services/index.ts
@@ -40,3 +40,4 @@ export * from "./contact.service";
 export * from "./media.service";
 export * from "./blog.service";
 export * from "./seller.service";
+export * from "./promotions.service";